AI-IntakeandAssessment/Questionnaire_agent/requirements.txt
azure-identity
azure-storage-blob
azure-search-documents
azure-cosmos
semantic-kernel
openpyxl
python-docx
pypdf
python-dotenv
azure-ai-projects